+++ argopy/tests/helpers/utils.py
@@ -19,7 +19,8 @@ from argopy.utilities import (
     list_available_index_src,
     isconnected,
     erddap_ds_exists,
-    isAPIconnected
+    isAPIconnected,
+    OceanOPSDeployments,
 )
 import logging
 
@@ -162,6 +163,13 @@ has_seaborn, requires_seaborn = _importorskip("seaborn")
 has_cartopy, requires_cartopy = _importorskip("cartopy")
 has_ipython, requires_ipython = _importorskip("IPython")
 
+#################
+# Ocean-OPS API #
+#################
+has_oops, requires_oops = _connectskip(
+    isconnected(OceanOPSDeployments().api_server_check), "a live Ocean-OPS server"
+)
+
 ############
 # Fix for issues discussed here:
 # - https://github.com/euroargodev/argopy/issues/63#issuecomment-742379699
@@ -237,6 +245,11 @@ def fct_safe_to_server_errors(func, *args, **kwargs):
             # ftplib.error_temp: 426 Failure writing network stream.
             msg = "\nCannot connect to the FTP server\n%s" % str(e.args)
             xmsg = "Failing because cannot connect to the FTP server, but should work"
+            pass
+        except fsspec.exceptions.FSTimeoutError as e:
+            # Sometimes, mostly from FTP, the time out is not long enough !            
+            msg = "\nUnexpected server time out\n%s" % str(e.args)
+            xmsg = "Failing because the server is temporarily too slow to respond, but should work"
             pass
         except Exception as e:
             msg = "\nUnknown server error:\n%s" % str(e.args)