+++ docs/whats-new.rst
@@ -13,7 +13,7 @@ v0.1.13 (XX)
 
 **Internals**
 
-- New method to check status of web API. Now allows for a keyword check rather than a simple url ping.
+- New method to check status of web API: now allows for a keyword check rather than a simple url ping. This comes with 2 new utilities functions :meth:`utilities.urlhaskeyword` and :meth:`utilities.isalive`. (:pr:`247`) by `G. Maze <http://www.github.com/gmaze>`_.
 
 
 v0.1.12 (16 May 2022)